requirements.txt
PyPDF2>=3.0.0
nltk>=3.8.1
python-magic>=0.4.27
numpy>=1.24.0
tqdm==4.66.1  # do paska postępu 